Retrieve committees or committee

Retrieve available committees, or a specific committee

committee(slug = NULL, ...)

Arguments

slug

An optional character string containing a "slug" referring to a political committee.

...

Arguments passed on to ppQuery, httr::GET

args

A character string containing additional URL-encoded parameters.

verbose

Whether to display informative messages on the query.

config

Additional configuration settings such as http authentication (authenticate()), additional headers (add_headers()), cookies (set_cookies()) etc. See config() for full details and list of helpers.

handle

The handle to use with this request. If not supplied, will be retrieved and reused from the handle_pool() based on the scheme, hostname and port of the url. By default httr requests to the same scheme/host/port combo. This substantially reduces connection time, and ensures that cookies are maintained over multiple requests to the same host. See handle_pool() for more details.

Value

The result of the query, typically a list object.

See also

Author

Thomas J. Leeper

Examples

# \dontrun{

committee("warren-for-senate-2012")
#> $committee
#> $committee$canonical_committee_id
#> NULL
#> 
#> $committee$fec_id
#> NULL
#> 
#> $committee$id
#> [1] 3934
#> 
#> $committee$name
#> [1] "WARREN FOR SENATE 2012"
#> 
#> $committee$pactrack
#> [1] FALSE
#> 
#> $committee$party
#> NULL
#> 
#> $committee$slug
#> [1] "warren-for-senate-2012"
#> 
#> $committee$state_id
#> NULL
#> 
#> $committee$freed_files
#>    filing.callsign filing.contract_number
#> 1           WBZ-TV                  58469
#> 2           WBZ-TV                  58398
#> 3           WBZ-TV           1061-58397-1
#> 4             WHDH                  48466
#> 5           WBZ-TV           1061-58409-1
#> 6           WBZ-TV           1061-58469-1
#> 7           WBZ-TV                  58191
#> 8             WHDH                  47708
#> 9             WHDH                  47708
#> 10            WFXT                 138032
#> 11          WBZ-TV                  58025
#> 12          WBZ-TV                  58120
#> 13          WBZ-TV                  58029
#> 14          WBZ-TV                  58126
#> 15          WBZ-TV           1061-58126-1
#> 16            WFXT                 122924
#> 17            WFXT                 126758
#> 18            WFXT                 130035
#> 19            WFXT             121348 / 1
#> 20            WFXT                 133284
#> 21            WFXT                 122656
#> 22            WFXT                 122900
#> 23            WFXT                 122911
#> 24            WFXT                 125362
#> 25            WFXT                 117346
#> 26            WHDH                  48479
#> 27            WHDH                  47874
#> 28            WHDH                  47708
#> 29            WHDH                  48263
#> 30            WHDH                  48293
#> 31          WBZ-TV           1061-58029-1
#> 32          WBZ-TV           1061-58190-1
#> 33          WBZ-TV           1061-58192-1
#> 34          WBZ-TV                  58190
#> 35          WBZ-TV                  58192
#> 36          WBZ-TV           1061-58191-1
#> 37          WBZ-TV           1061-58244-1
#> 38          WBZ-TV                  58409
#> 39          WBZ-TV                  58397
#> 40          WBZ-TV                  58244
#> 41          WBZ-TV           1061-58398-1
#> 42            WHDH                  48127
#> 43          WBZ-TV                  58301
#> 44          WBZ-TV           1061-58301-1
#> 45            WFXT                 127543
#> 46            WFXT                 129040
#> 47            WFXT                 129089
#> 48            WHDH                  48262
#> 49            WFXT                 117320
#> 50            WHDH                  48048
#> 51            WFXT                 121348
#> 52            WFXT                 132261
#> 53            WFXT                 135587
#> 54            WHDH                  47702
#> 55            WHDH                  47711
#> 56          WBZ-TV           1061-58025-1
#> 57          WBZ-TV           1061-58120-1
#>                                            filing.dc_slug filing.filing_type
#> 1  444353-collect-files-25456-political-file-2012-federal            Federal
#> 2  443225-collect-files-25456-political-file-2012-federal            Federal
#> 3  443222-collect-files-25456-political-file-2012-federal            Federal
#> 4  444345-collect-files-72145-political-file-2012-federal            Federal
#> 5  443226-collect-files-25456-political-file-2012-federal            Federal
#> 6  444352-collect-files-25456-political-file-2012-federal            Federal
#> 7  431719-collect-files-25456-political-file-2012-federal            Federal
#> 8   477716-warren-for-senate-est-11666-new-10-22-10-28-12            Federal
#> 9  477717-warren-for-senate-est-11666-revised-10-22-10-28            Federal
#> 10           476419-warren-est-12570-13505067103162-_-pdf            Federal
#> 11 418107-collect-files-25456-political-file-2012-federal            Federal
#> 12 418112-collect-files-25456-political-file-2012-federal            Federal
#> 13 418110-collect-files-25456-political-file-2012-federal            Federal
#> 14 418114-collect-files-25456-political-file-2012-federal            Federal
#> 15 418113-collect-files-25456-political-file-2012-federal            Federal
#> 16  427463-collect-files-6463-political-file-2012-federal            Federal
#> 17           455922-warren-est-11954-13494593967715-_-pdf            Federal
#> 18  442846-collect-files-6463-political-file-2012-federal            Federal
#> 19  421303-collect-files-6463-political-file-2012-federal            Federal
#> 20           455923-warren-est-12217-13493802078464-_-pdf            Federal
#> 21  422313-collect-files-6463-political-file-2012-federal            Federal
#> 22  427461-collect-files-6463-political-file-2012-federal            Federal
#> 23  427462-collect-files-6463-political-file-2012-federal            Federal
#> 24  427464-collect-files-6463-political-file-2012-federal            Federal
#> 25  431336-collect-files-6463-political-file-2012-federal            Federal
#> 26 444346-collect-files-72145-political-file-2012-federal            Federal
#> 27 439538-collect-files-72145-political-file-2012-federal            Federal
#> 28 421178-collect-files-72145-political-file-2012-federal            Federal
#> 29 439541-collect-files-72145-political-file-2012-federal            Federal
#> 30 443211-collect-files-72145-political-file-2012-federal            Federal
#> 31 418108-collect-files-25456-political-file-2012-federal            Federal
#> 32 424223-collect-files-25456-political-file-2012-federal            Federal
#> 33 424225-collect-files-25456-political-file-2012-federal            Federal
#> 34 431718-collect-files-25456-political-file-2012-federal            Federal
#> 35 431720-collect-files-25456-political-file-2012-federal            Federal
#> 36 424224-collect-files-25456-political-file-2012-federal            Federal
#> 37 427579-collect-files-25456-political-file-2012-federal            Federal
#> 38 443227-collect-files-25456-political-file-2012-federal            Federal
#> 39 443223-collect-files-25456-political-file-2012-federal            Federal
#> 40 427580-collect-files-25456-political-file-2012-federal            Federal
#> 41 443224-collect-files-25456-political-file-2012-federal            Federal
#> 42 439539-collect-files-72145-political-file-2012-federal            Federal
#> 43 434122-collect-files-25456-political-file-2012-federal            Federal
#> 44 434121-collect-files-25456-political-file-2012-federal            Federal
#> 45  434989-collect-files-6463-political-file-2012-federal            Federal
#> 46  439604-collect-files-6463-political-file-2012-federal            Federal
#> 47  439605-collect-files-6463-political-file-2012-federal            Federal
#> 48 439540-collect-files-72145-political-file-2012-federal            Federal
#> 49  431335-collect-files-6463-political-file-2012-federal            Federal
#> 50 431257-collect-files-72145-political-file-2012-federal            Federal
#> 51  417074-collect-files-6463-political-file-2012-federal            Federal
#> 52           452965-warren-est-12184-13492788108564-_-pdf            Federal
#> 53           462051-warren-est-12228-13498878560801-_-pdf            Federal
#> 54 421177-collect-files-72145-political-file-2012-federal            Federal
#> 55 421179-collect-files-72145-political-file-2012-federal            Federal
#> 56 418105-collect-files-25456-political-file-2012-federal            Federal
#> 57 418111-collect-files-25456-political-file-2012-federal            Federal
#>    filing.gross_amount filing.id filing.market_id
#> 1               355510     17693                7
#> 2               212570     16868                7
#> 3                21440     16865                7
#> 4               323500     17685                7
#> 5                34430     16869                7
#> 6               355510     17692                7
#> 7               359550     11774                7
#> 8               325900     37293                7
#> 9               405350     37294                7
#> 10               16200     35998                7
#> 11               79104      3912                7
#> 12              106865      3917                7
#> 13               95130      3915                7
#> 14              122565      3919                7
#> 15              109070      3918                7
#> 16              137150      8940                7
#> 17              143450     20955                7
#> 18               28000     16538                7
#> 19               82850      6769                7
#> 20               13000     20956                7
#> 21               82725      7454                7
#> 22              101150      8938                7
#> 23              146250      8939                7
#> 24              164260      8941                7
#> 25               49425     11396                7
#> 26               20000     17686                7
#> 27              204200     15396                7
#> 28              325900      6644                7
#> 29               35000     15399                7
#> 30               69300     16855                7
#> 31               91440      3913                7
#> 32              286060      7690                7
#> 33              330495      7692                7
#> 34              297940     11773                7
#> 35              330495     11775                7
#> 36              359550      7691                7
#> 37              196955      9056                7
#> 38               34430     16870                7
#> 39               21440     16866                7
#> 40              196955      9057                7
#> 41              212570     16867                7
#> 42               53900     15397                7
#> 43              236635     13306                7
#> 44              234835     13305                7
#> 45               54400     13781                7
#> 46              127050     15460                7
#> 47               14400     15461                7
#> 48              200650     15398                7
#> 49               52375     11395                7
#> 50              264850     11318                7
#> 51               82850      2904                7
#> 52              184875     18888                7
#> 53              192500     24899                7
#> 54              293600      6643                7
#> 55              394650      6645                7
#> 56               77704      3910                7
#> 57              103175      3916                7
#>                                                                                                                     filing.thumbnail_url
#> 1  https://s3.amazonaws.com/s3.documentcloud.org/documents/444353/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 2  https://s3.amazonaws.com/s3.documentcloud.org/documents/443225/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 3  https://s3.amazonaws.com/s3.documentcloud.org/documents/443222/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 4  https://s3.amazonaws.com/s3.documentcloud.org/documents/444345/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 5  https://s3.amazonaws.com/s3.documentcloud.org/documents/443226/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 6  https://s3.amazonaws.com/s3.documentcloud.org/documents/444352/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 7  https://s3.amazonaws.com/s3.documentcloud.org/documents/431719/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 8   https://s3.amazonaws.com/s3.documentcloud.org/documents/477716/pages/warren-for-senate-est-11666-new-10-22-10-28-12-p1-thumbnail.gif
#> 9  https://s3.amazonaws.com/s3.documentcloud.org/documents/477717/pages/warren-for-senate-est-11666-revised-10-22-10-28-p1-thumbnail.gif
#> 10           https://s3.amazonaws.com/s3.documentcloud.org/documents/476419/pages/warren-est-12570-13505067103162-_-pdf-p1-thumbnail.gif
#> 11 https://s3.amazonaws.com/s3.documentcloud.org/documents/418107/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 12 https://s3.amazonaws.com/s3.documentcloud.org/documents/418112/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 13 https://s3.amazonaws.com/s3.documentcloud.org/documents/418110/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 14 https://s3.amazonaws.com/s3.documentcloud.org/documents/418114/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 15 https://s3.amazonaws.com/s3.documentcloud.org/documents/418113/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 16  https://s3.amazonaws.com/s3.documentcloud.org/documents/427463/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 17           https://s3.amazonaws.com/s3.documentcloud.org/documents/455922/pages/warren-est-11954-13494593967715-_-pdf-p1-thumbnail.gif
#> 18  https://s3.amazonaws.com/s3.documentcloud.org/documents/442846/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 19  https://s3.amazonaws.com/s3.documentcloud.org/documents/421303/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 20           https://s3.amazonaws.com/s3.documentcloud.org/documents/455923/pages/warren-est-12217-13493802078464-_-pdf-p1-thumbnail.gif
#> 21  https://s3.amazonaws.com/s3.documentcloud.org/documents/422313/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 22  https://s3.amazonaws.com/s3.documentcloud.org/documents/427461/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 23  https://s3.amazonaws.com/s3.documentcloud.org/documents/427462/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 24  https://s3.amazonaws.com/s3.documentcloud.org/documents/427464/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 25  https://s3.amazonaws.com/s3.documentcloud.org/documents/431336/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 26 https://s3.amazonaws.com/s3.documentcloud.org/documents/444346/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 27 https://s3.amazonaws.com/s3.documentcloud.org/documents/439538/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 28 https://s3.amazonaws.com/s3.documentcloud.org/documents/421178/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 29 https://s3.amazonaws.com/s3.documentcloud.org/documents/439541/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 30 https://s3.amazonaws.com/s3.documentcloud.org/documents/443211/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 31 https://s3.amazonaws.com/s3.documentcloud.org/documents/418108/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 32 https://s3.amazonaws.com/s3.documentcloud.org/documents/424223/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 33 https://s3.amazonaws.com/s3.documentcloud.org/documents/424225/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 34 https://s3.amazonaws.com/s3.documentcloud.org/documents/431718/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 35 https://s3.amazonaws.com/s3.documentcloud.org/documents/431720/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 36 https://s3.amazonaws.com/s3.documentcloud.org/documents/424224/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 37 https://s3.amazonaws.com/s3.documentcloud.org/documents/427579/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 38 https://s3.amazonaws.com/s3.documentcloud.org/documents/443227/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 39 https://s3.amazonaws.com/s3.documentcloud.org/documents/443223/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 40 https://s3.amazonaws.com/s3.documentcloud.org/documents/427580/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 41 https://s3.amazonaws.com/s3.documentcloud.org/documents/443224/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 42 https://s3.amazonaws.com/s3.documentcloud.org/documents/439539/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 43 https://s3.amazonaws.com/s3.documentcloud.org/documents/434122/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 44 https://s3.amazonaws.com/s3.documentcloud.org/documents/434121/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 45  https://s3.amazonaws.com/s3.documentcloud.org/documents/434989/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 46  https://s3.amazonaws.com/s3.documentcloud.org/documents/439604/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 47  https://s3.amazonaws.com/s3.documentcloud.org/documents/439605/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 48 https://s3.amazonaws.com/s3.documentcloud.org/documents/439540/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 49  https://s3.amazonaws.com/s3.documentcloud.org/documents/431335/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 50 https://s3.amazonaws.com/s3.documentcloud.org/documents/431257/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 51  https://s3.amazonaws.com/s3.documentcloud.org/documents/417074/pages/collect-files-6463-political-file-2012-federal-p1-thumbnail.gif
#> 52           https://s3.amazonaws.com/s3.documentcloud.org/documents/452965/pages/warren-est-12184-13492788108564-_-pdf-p1-thumbnail.gif
#> 53           https://s3.amazonaws.com/s3.documentcloud.org/documents/462051/pages/warren-est-12228-13498878560801-_-pdf-p1-thumbnail.gif
#> 54 https://s3.amazonaws.com/s3.documentcloud.org/documents/421177/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 55 https://s3.amazonaws.com/s3.documentcloud.org/documents/421179/pages/collect-files-72145-political-file-2012-federal-p1-thumbnail.gif
#> 56 https://s3.amazonaws.com/s3.documentcloud.org/documents/418105/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#> 57 https://s3.amazonaws.com/s3.documentcloud.org/documents/418111/pages/collect-files-25456-political-file-2012-federal-p1-thumbnail.gif
#>    filing.transcriptions_count   filing.upload_date
#> 1                            2 2012-09-28T00:00:00Z
#> 2                            2 2012-09-25T00:00:00Z
#> 3                            3 2012-09-25T00:00:00Z
#> 4                            3 2012-09-28T00:00:00Z
#> 5                            3 2012-09-25T00:00:00Z
#> 6                            3 2012-09-28T00:00:00Z
#> 7                            2 2012-09-05T00:00:00Z
#> 8                            3 2012-10-18T00:00:00Z
#> 9                            3 2012-10-18T00:00:00Z
#> 10                           2 2012-10-17T00:00:00Z
#> 11                           2 2012-08-14T00:00:00Z
#> 12                           2 2012-08-23T00:00:00Z
#> 13                           2 2012-08-14T00:00:00Z
#> 14                           2 2012-08-23T00:00:00Z
#> 15                           4 2012-08-27T00:00:00Z
#> 16                           2 2012-09-11T00:00:00Z
#> 17                           3 2012-10-05T00:00:00Z
#> 18                           2 2012-09-25T00:00:00Z
#> 19                           3 2012-08-31T00:00:00Z
#> 20                           2 2012-10-04T00:00:00Z
#> 21                           3 2012-08-31T00:00:00Z
#> 22                           2 2012-09-11T00:00:00Z
#> 23                           2 2012-09-11T00:00:00Z
#> 24                           2 2012-09-11T00:00:00Z
#> 25                           2 2012-08-10T00:00:00Z
#> 26                           3 2012-09-28T00:00:00Z
#> 27                           4 2012-09-21T00:00:00Z
#> 28                           3 2012-08-31T00:00:00Z
#> 29                           4 2012-09-21T00:00:00Z
#> 30                           3 2012-09-24T00:00:00Z
#> 31                           4 2012-08-14T00:00:00Z
#> 32                           3 2012-09-05T00:00:00Z
#> 33                           3 2012-09-05T00:00:00Z
#> 34                           3 2012-09-05T00:00:00Z
#> 35                           2 2012-09-05T00:00:00Z
#> 36                           3 2012-09-05T00:00:00Z
#> 37                           2 2012-09-10T00:00:00Z
#> 38                           2 2012-09-25T00:00:00Z
#> 39                           2 2012-09-25T00:00:00Z
#> 40                           2 2012-09-10T00:00:00Z
#> 41                           3 2012-09-25T00:00:00Z
#> 42                           3 2012-09-21T00:00:00Z
#> 43                           3 2012-09-14T00:00:00Z
#> 44                           2 2012-09-14T00:00:00Z
#> 45                           2 2012-09-17T00:00:00Z
#> 46                           2 2012-09-21T00:00:00Z
#> 47                           2 2012-09-21T00:00:00Z
#> 48                           2 2012-09-21T00:00:00Z
#> 49                           2 2012-08-10T00:00:00Z
#> 50                           2 2012-09-13T00:00:00Z
#> 51                           2 2012-08-24T00:00:00Z
#> 52                           2 2012-10-03T00:00:00Z
#> 53                           3 2012-10-10T00:00:00Z
#> 54                           3 2012-08-31T00:00:00Z
#> 55                           3 2012-08-31T00:00:00Z
#> 56                           5 2012-08-14T00:00:00Z
#> 57                           4 2012-08-27T00:00:00Z
#>                                                                                                                                                   filing.url
#> 1      /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58469 FED ORDER (13488447324801)_.pdf
#> 2      /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58398 FED ORDER (13485819033698)_.pdf
#> 3      /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58397 FED TRACT (13485819065896)_.pdf
#> 4        /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 12184 NEW 10.1.12 - 10.7.12 (13488454296047)_.pdf
#> 5      /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58409 FED TRACT (13485905719417)_.pdf
#> 6      /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58469 FED TRACT (13488447456255)_.pdf
#> 7      /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58191 FED ORDER (13468551332360)_.pdf
#> 8           /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 11666 NEW 10.22-10.28.12 (13505761847846)_.pdf
#> 9       /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 11666 Revised 10.22-10.28.12 (13505761848354)_.pdf
#> 10                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 12570 (13505067103162)_.pdf
#> 11     /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58025 FED ORDER (13449696816301)_.pdf
#> 12    /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ WARREN FOR SENATE D SENATEMA ORD58120 FED ORDER (13457355056808)_.pdf
#> 13     /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58029 FED ORDER (13449696816990)_.pdf
#> 14    /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ WARREN FOR SENATE D SENATEMA ORD58126 FED ORDER (13457355068812)_.pdf
#> 15 /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ WARREN FOR SENATE D SENATEMA ORD58126 FED CONTRACT (13460805058183)_.pdf
#> 16                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 11667 (13473746774509)_.pdf
#> 17                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 11954 (13494593967715)_.pdf
#> 18                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 12156 (13485795067282)_.pdf
#> 19                                       /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/WARREN EST 11662.1 (13464534032681)_.pdf
#> 20                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 12217 (13493802078464)_.pdf
#> 21                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/WARREN EST 11664 (13464543124336)_.pdf
#> 22                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 11665 (13473748727823)_.pdf
#> 23                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 11666 (13473746774224)_.pdf
#> 24                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 11862 (13473746774917)_.pdf
#> 25                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 11586 (13446273164727)_.pdf
#> 26       /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 12217 NEW 10.1.12 - 10.7.12 (13488459399306)_.pdf
#> 27             /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 11862 REV 1 9.10-9.16 (13482567141112)_.pdf
#> 28          /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 11666 NEW 10.22-10.28.12 (13464254675922)_.pdf
#> 29             /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 12100 9.24.12-9.30.12 (13482606128095)_.pdf
#> 30       /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 12156 NEW 9.25.12 - 9.30.12 (13485201542678)_.pdf
#> 31  /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58029 FED CONTRACT (13449777167988)_.pdf
#> 32  /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58190 FED CONTRACT (13468779206050)_.pdf
#> 33  /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58192 FED CONTRACT (13468776145271)_.pdf
#> 34     /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58190 FED ORDER (13468551364772)_.pdf
#> 35     /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58192 FED ORDER (13468551321346)_.pdf
#> 36  /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58191 FED CONTRACT (13468776134107)_.pdf
#> 37  /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58244 FED CONTRACT (13472946211177)_.pdf
#> 38     /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58409 FED ORDER (13485905709846)_.pdf
#> 39     /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58397 FED ORDER (13485819044576)_.pdf
#> 40     /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58244 FED ORDER (13472946200060)_.pdf
#> 41     /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58398 FED TRACT (13485819055209)_.pdf
#> 42         /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 12054 New 9.18.12-9.23.12 (13482594068604)_.pdf
#> 43     /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58301 FED ORDER (13476582100669)_.pdf
#> 44  /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58301 FED CONTRACT (13476582111538)_.pdf
#> 45                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 12054 (13479165060321)_.pdf
#> 46                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 12099 (13482405064495)_.pdf
#> 47                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 12100 (13482405054072)_.pdf
#> 48             /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 12099 9.24.12-9.30.12 (13482606084640)_.pdf
#> 49                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 11585 (13446272252913)_.pdf
#> 50               /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 11954 NEW 9.17-9.23 (13475484124015)_.pdf
#> 51                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 11662 (13458400872787)_.pdf
#> 52                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 12184 (13492788108564)_.pdf
#> 53                                         /collect/files/6463/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren Est 12228 (13498878560801)_.pdf
#> 54          /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 11665 NEW 10.15-10.21.12 (13464254674972)_.pdf
#> 55           /collect/files/72145/Political File/2012/Federal/US Senate/Elizabeth Warren/Warren for Senate Est 11667 NEW 10.29-11.6.12 (13464254677157)_.pdf
#> 56  /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58025 FED CONTRACT (13449771554387)_.pdf
#> 57  /collect/files/25456/Political File/2012/Federal/US Senate/ELIZABETH WARREN /WBZ ELIZABETH WARREN D SENATEMA ORD58120 FED CONTRACT (13460805069185)_.pdf
#>    filing.committee.id  filing.committee.name  filing.committee.slug
#> 1                 3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 2                 3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 3                 3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 4                 3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 5                 3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 6                 3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 7                 3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 8                 3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 9                 3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 10                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 11                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 12                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 13                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 14                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 15                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 16                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 17                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 18                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 19                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 20                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 21                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 22                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 23                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 24                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 25                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 26                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 27                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 28                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 29                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 30                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 31                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 32                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 33                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 34                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 35                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 36                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 37                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 38                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 39                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 40                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 41                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 42                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 43                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 44                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 45                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 46                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 47                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 48                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 49                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 50                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 51                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 52                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 53                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 54                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 55                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 56                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#> 57                3934 WARREN FOR SENATE 2012 warren-for-senate-2012
#>             filing.agency.name
#> 1  MEDIA STRATEGIES & RESEARCH
#> 2  MEDIA STRATEGIES & RESEARCH
#> 3  MEDIA STRATEGIES & RESEARCH
#> 4  MEDIA STRATEGIES & RESEARCH
#> 5  MEDIA STRATEGIES & RESEARCH
#> 6  MEDIA STRATEGIES & RESEARCH
#> 7  MEDIA STRATEGIES & RESEARCH
#> 8  MEDIA STRATEGIES & RESEARCH
#> 9  MEDIA STRATEGIES & RESEARCH
#> 10 MEDIA STRATEGIES & RESEARCH
#> 11 MEDIA STRATEGIES & RESEARCH
#> 12 MEDIA STRATEGIES & RESEARCH
#> 13 MEDIA STRATEGIES & RESEARCH
#> 14 MEDIA STRATEGIES & RESEARCH
#> 15 MEDIA STRATEGIES & RESEARCH
#> 16 MEDIA STRATEGIES & RESEARCH
#> 17 MEDIA STRATEGIES & RESEARCH
#> 18 MEDIA STRATEGIES & RESEARCH
#> 19 MEDIA STRATEGIES & RESEARCH
#> 20 MEDIA STRATEGIES & RESEARCH
#> 21 MEDIA STRATEGIES & RESEARCH
#> 22 MEDIA STRATEGIES & RESEARCH
#> 23 MEDIA STRATEGIES & RESEARCH
#> 24 MEDIA STRATEGIES & RESEARCH
#> 25 MEDIA STRATEGIES & RESEARCH
#> 26 MEDIA STRATEGIES & RESEARCH
#> 27 MEDIA STRATEGIES & RESEARCH
#> 28 MEDIA STRATEGIES & RESEARCH
#> 29 MEDIA STRATEGIES & RESEARCH
#> 30 MEDIA STRATEGIES & RESEARCH
#> 31 MEDIA STRATEGIES & RESEARCH
#> 32 MEDIA STRATEGIES & RESEARCH
#> 33 MEDIA STRATEGIES & RESEARCH
#> 34 MEDIA STRATEGIES & RESEARCH
#> 35 MEDIA STRATEGIES & RESEARCH
#> 36 MEDIA STRATEGIES & RESEARCH
#> 37 MEDIA STRATEGIES & RESEARCH
#> 38 MEDIA STRATEGIES & RESEARCH
#> 39 MEDIA STRATEGIES & RESEARCH
#> 40 MEDIA STRATEGIES & RESEARCH
#> 41 MEDIA STRATEGIES & RESEARCH
#> 42 MEDIA STRATEGIES & RESEARCH
#> 43 MEDIA STRATEGIES & RESEARCH
#> 44 MEDIA STRATEGIES & RESEARCH
#> 45 MEDIA STRATEGIES & RESEARCH
#> 46 MEDIA STRATEGIES & RESEARCH
#> 47 MEDIA STRATEGIES & RESEARCH
#> 48 MEDIA STRATEGIES & RESEARCH
#> 49 MEDIA STRATEGIES & RESEARCH
#> 50 MEDIA STRATEGIES & RESEARCH
#> 51 MEDIA STRATEGIES & RESEARCH
#> 52 MEDIA STRATEGIES & RESEARCH
#> 53 MEDIA STRATEGIES & RESEARCH
#> 54 MEDIA STRATEGIES & RESEARCH
#> 55 MEDIA STRATEGIES & RESEARCH
#> 56 MEDIA STRATEGIES & RESEARCH
#> 57 MEDIA STRATEGIES & RESEARCH
#> 
#> $committee$freed_ct
#> [1] 57
#> 
#> $committee$gross_amount
#> [1] 9452133
#> 
#> $committee$from_pactrack
#> NULL
#> 
#> 
# }