Retrieve filing

Retrieve a specific filing, by ID

filing(id, ...)

Arguments

id

A character string containing a filing identification number.

...

Arguments passed on to ppQuery, httr::GET

args

A character string containing additional URL-encoded parameters.

verbose

Whether to display informative messages on the query.

config

Additional configuration settings such as http authentication (authenticate()), additional headers (add_headers()), cookies (set_cookies()) etc. See config() for full details and list of helpers.

handle

The handle to use with this request. If not supplied, will be retrieved and reused from the handle_pool() based on the scheme, hostname and port of the url. By default httr requests to the same scheme/host/port combo. This substantially reduces connection time, and ensures that cookies are maintained over multiple requests to the same host. See handle_pool() for more details.

Value

The result of the query, typically a list object.

See also

Author

Thomas J. Leeper

Examples

# \dontrun{

filing(51212)
#> $callsign
#> [1] "WHP-TV"
#> 
#> $contract_number
#> [1] "69338"
#> 
#> $dc_slug
#> [1] "501618-69338-2-13519611148107-_-pdf"
#> 
#> $filing_type
#> [1] "Federal"
#> 
#> $gross_amount
#> [1] 35940
#> 
#> $id
#> [1] 51212
#> 
#> $market_id
#> [1] 41
#> 
#> $thumbnail_url
#> [1] "https://s3.amazonaws.com/s3.documentcloud.org/documents/501618/pages/69338-2-13519611148107-_-pdf-p1-thumbnail.gif"
#> 
#> $transcriptions_count
#> [1] 2
#> 
#> $upload_date
#> [1] "2012-11-03T00:00:00Z"
#> 
#> $url
#> [1] "/collect/files/72313/Political File/2012/Federal/President/BARACK OBAMA/69338/69338-2 (13519611148107)_.pdf"
#> 
#> $committee
#> $committee$id
#> [1] 1221
#> 
#> $committee$name
#> [1] "OBAMA FOR AMERICA"
#> 
#> $committee$slug
#> [1] "obama-for-america"
#> 
#> 
#> $agency
#> $agency$name
#> [1] "GREER MARGOLIS MITCHELL"
#> 
#> 
# }