Forensics API

Retrieve details about a given state or system

geos(state = NULL, ...)

systems(id, ...)

Arguments

state

An optional character string containing a two-letter state postal code. If NULL, results for all available states are returned.

...

Arguments passed on to ppQuery, httr::GET

args

A character string containing additional URL-encoded parameters.

verbose

Whether to display informative messages on the query.

config

Additional configuration settings such as http authentication (authenticate()), additional headers (add_headers()), cookies (set_cookies()) etc. See config() for full details and list of helpers.

handle

The handle to use with this request. If not supplied, will be retrieved and reused from the handle_pool() based on the scheme, hostname and port of the url. By default httr requests to the same scheme/host/port combo. This substantially reduces connection time, and ensures that cookies are maintained over multiple requests to the same host. See handle_pool() for more details.

id

The identification number for a particular medical examiner or coroner system.

Value

The result of the query, typically a list object.

References

API Documentation

Author

Thomas J. Leeper

Examples

# \dontrun{

geos("48029")
#> $geo
#> $geo$cached_slug
#> [1] "bexar-county"
#> 
#> $geo$coroner_system
#> [1] "none"
#> 
#> $geo$created_at
#> [1] "2011-05-04T12:23:12-04:00"
#> 
#> $geo$district
#> NULL
#> 
#> $geo$id
#> [1] 48029
#> 
#> $geo$kind
#> [1] "county"
#> 
#> $geo$medical_examiner_system
#> [1] "county"
#> 
#> $geo$name
#> [1] "Bexar County"
#> 
#> $geo$note
#> NULL
#> 
#> $geo$population
#> NULL
#> 
#> $geo$postal_abbreviation
#> [1] "TX"
#> 
#> $geo$updated_at
#> [1] "2011-05-04T12:24:56-04:00"
#> 
#> $geo$system_types
#> $geo$system_types$medical_examiner_system
#> [1] "county"
#> 
#> $geo$system_types$coroner_system
#> [1] "none"
#> 
#> 
#> $geo$medical_examiners
#>   accreditation autopsies bc_forensic_pathologists cached_slug
#> 1          Full        NA                        5          NA
#>                  created_at fellows forensic_pathologists id
#> 1 2011-05-04T12:23:52-04:00       1                    NA 34
#>                                    name note pending
#> 1 Bexar County (Texas) Medical Examiner   NA      NA
#>   pt_bc_forensic_pathologists pt_uc_forensic_pathologists
#> 1                           0                           0
#>   uc_forensic_pathologists                updated_at
#> 1                        0 2011-05-04T12:23:53-04:00
#>                                 website
#> 1 http://www.bexar.org/medicalexaminer/
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          stats
#> 1 population, pt_provisional_forensic_pathologists, autopsies, annual_fiscal_budget, forensic_pathologists, contract_cases, population, pt_provisional_forensic_pathologists, autopsies, annual_fiscal_budget, forensic_pathologists, contract_cases, population, pt_provisional_forensic_pathologists, autopsies, annual_fiscal_budget, forensic_pathologists, contract_cases, population, pt_provisional_forensic_pathologists, autopsies, annual_fiscal_budget, forensic_pathologists, contract_cases, population, pt_provisional_forensic_pathologists, autopsies, annual_fiscal_budget, forensic_pathologists, contract_cases, population, pt_provisional_forensic_pathologists, autopsies, annual_fiscal_budget, forensic_pathologists, contract_cases, population, pt_provisional_forensic_pathologists, annual_fiscal_budget, forensic_pathologists, autopsy_ratio, raw_autopsy_rate, autopsy_standard_deviation, autopsy_ratio, raw_autopsy_rate, autopsy_standard_deviation, autopsy_ratio, raw_autopsy_rate, autopsy_standard_deviation, autopsy_ratio, raw_autopsy_rate, autopsy_standard_deviation, 1487460, 0, 1300, 0, 5, 241, 1511510, 0, 1247, 2979300, 5, 243, 1550920, 0, 1323, 3133380, 4, 220, 1588900, 0, 1348, 3250820, 4, 118, 1621300, 0, 1303, 3690210, 4, 76, 1651450, 0, 1221, 3972260, 5, 71, 0, 0, 4041600, 5, -0.10843, 9.5184, -0.39, -0.0654056, 10.1282, -0.24, -0.0256922, 10.3452, -0.09, 0.0866295, 11.4835, 0.31, 2004, 2004, 2004, 2004, 2004, 2004, 2005, 2005, 2005, 2005, 2005, 2005, 2006, 2006, 2006, 2006, 2006, 2006, 2007, 2007, 2007, 2007, 2007, 2007, 2008, 2008, 2008, 2008, 2008, 2008, 2009, 2009, 2009, 2009, 2009, 2009, 2010, 2010, 2010, 2010, 2005, 2005, 2005, 2004, 2004, 2004, 2006, 2006, 2006, 2007, 2007, 2007
#> 
#> $geo$coroners
#> list()
#> 
#> 

systems(39)
#> $accreditation
#> [1] ""
#> 
#> $autopsies
#> NULL
#> 
#> $bc_forensic_pathologists
#> [1] 2
#> 
#> $cached_slug
#> NULL
#> 
#> $created_at
#> [1] "2011-05-04T12:24:00-04:00"
#> 
#> $fellows
#> [1] 0
#> 
#> $forensic_pathologists
#> NULL
#> 
#> $id
#> [1] 39
#> 
#> $name
#> [1] "Erie County (N.Y.) Medical Examiner"
#> 
#> $note
#> NULL
#> 
#> $pending
#> NULL
#> 
#> $pt_bc_forensic_pathologists
#> [1] 1
#> 
#> $pt_uc_forensic_pathologists
#> [1] 0
#> 
#> $uc_forensic_pathologists
#> [1] 1
#> 
#> $updated_at
#> [1] "2011-05-04T12:24:00-04:00"
#> 
#> $website
#> [1] "http://www2.erie.gov/health/index.php?q=medical-examiner"
#> 
#> $stats
#>                                    kind        value year
#> 1                            population  9.32002e+05 2004
#> 2  pt_provisional_forensic_pathologists  1.00000e+00 2004
#> 3                             autopsies  1.19800e+03 2004
#> 4                  annual_fiscal_budget  2.14476e+06 2004
#> 5                 forensic_pathologists  3.00000e+00 2004
#> 6                        contract_cases  2.76000e+02 2004
#> 7                            population  9.23820e+05 2005
#> 8  pt_provisional_forensic_pathologists  1.00000e+00 2005
#> 9                             autopsies  1.03900e+03 2005
#> 10                 annual_fiscal_budget  2.03764e+06 2005
#> 11                forensic_pathologists  3.00000e+00 2005
#> 12                       contract_cases  2.40000e+02 2005
#> 13                           population  9.16292e+05 2006
#> 14 pt_provisional_forensic_pathologists  1.00000e+00 2006
#> 15                            autopsies  1.01600e+03 2006
#> 16                 annual_fiscal_budget  1.93932e+06 2006
#> 17                forensic_pathologists  3.00000e+00 2006
#> 18                       contract_cases  2.67000e+02 2006
#> 19                           population  9.11784e+05 2007
#> 20 pt_provisional_forensic_pathologists  1.00000e+00 2007
#> 21                            autopsies  9.21000e+02 2007
#> 22                 annual_fiscal_budget  2.05244e+06 2007
#> 23                forensic_pathologists  3.00000e+00 2007
#> 24                       contract_cases  2.48000e+02 2007
#> 25                           population  9.09858e+05 2008
#> 26 pt_provisional_forensic_pathologists  1.00000e+00 2008
#> 27                            autopsies  7.37000e+02 2008
#> 28                 annual_fiscal_budget  2.27608e+06 2008
#> 29                forensic_pathologists  3.00000e+00 2008
#> 30                       contract_cases  2.24000e+02 2008
#> 31                           population  9.09247e+05 2009
#> 32 pt_provisional_forensic_pathologists  1.00000e+00 2009
#> 33                            autopsies  6.19000e+02 2009
#> 34                 annual_fiscal_budget  2.92444e+06 2009
#> 35                forensic_pathologists  3.00000e+00 2009
#> 36                       contract_cases  1.62000e+02 2009
#> 37                           population  0.00000e+00 2010
#> 38 pt_provisional_forensic_pathologists  1.00000e+00 2010
#> 39                            autopsies  6.14000e+02 2010
#> 40                 annual_fiscal_budget  2.99839e+06 2010
#> 41                forensic_pathologists  3.00000e+00 2010
#> 42                       contract_cases  1.51000e+02 2010
#> 43                        autopsy_ratio -1.26467e-01 2007
#> 44                     raw_autopsy_rate  7.07300e+00 2007
#> 45           autopsy_standard_deviation -3.50000e-01 2007
#> 46                        autopsy_ratio  6.16924e-03 2006
#> 47                     raw_autopsy_rate  7.94270e+00 2006
#> 48           autopsy_standard_deviation  2.00000e-02 2006
#> 49                        autopsy_ratio  6.33676e-02 2005
#> 50                     raw_autopsy_rate  8.16560e+00 2005
#> 51           autopsy_standard_deviation  1.70000e-01 2005
#> 52                        autopsy_ratio  2.13259e-01 2004
#> 53                     raw_autopsy_rate  9.50710e+00 2004
#> 54           autopsy_standard_deviation  5.70000e-01 2004
#> 
#> $geos
#>   cached_slug coroner_system                created_at district    id   kind
#> 1 erie-county           none 2011-05-04T12:22:56-04:00       NA 36029 county
#>   medical_examiner_system        name note population postal_abbreviation
#> 1                  county Erie County   NA         NA                  NY
#>                  updated_at
#> 1 2011-05-04T12:24:57-04:00
#> 
# }