Nonprofit Explorer

Search for and retrieve nonprofit data

np_search(
  q = NULL,
  order = NULL,
  sort = NULL,
  state = NULL,
  ntee = NULL,
  subsection = NULL,
  ...
)

np_org(ein, ...)

Arguments

q

A character string containing a search query.

order

A character string containing a field by which to order the results. One of "name", "city", "state", "c_code", "fiscalyr", "revenue", "expenses", "assets", "liabilities", "pct_officer_compt". The ascending or descending sort by this parameter is controlled by sort.

sort

One of "asc" or "desc", controlling the order of the results (according the the ordering given in order).

state

A two-letter state postal abbreviation code for a U.S. state or territory. Use "ZZ" for foreign organizations.

ntee

A numeric value (between 1 and 10) containing a National Taxonomy of Exempt Entities (NTEE) Major Group code.

subsection

A numeric value (between 2 and 92) referring to a subsection of tax code section 501(c).

...

Arguments passed on to ppQuery, httr::GET

args

A character string containing additional URL-encoded parameters.

verbose

Whether to display informative messages on the query.

config

Additional configuration settings such as http authentication (authenticate()), additional headers (add_headers()), cookies (set_cookies()) etc. See config() for full details and list of helpers.

handle

The handle to use with this request. If not supplied, will be retrieved and reused from the handle_pool() based on the scheme, hostname and port of the url. By default httr requests to the same scheme/host/port combo. This substantially reduces connection time, and ensures that cookies are maintained over multiple requests to the same host. See handle_pool() for more details.

ein

An employer identification number (EIN).

Value

The result of the query, typically a list object.

Author

Thomas J. Leeper

Examples

# \dontrun{
np_search("propublica")
#> $total_results
#> [1] 1
#> 
#> $organizations
#>         ein     strein            name        sub_name     city state ntee_code
#> 1 142007220 14-2007220 PRO PUBLICA INC PRO PUBLICA INC NEW YORK    NY       A20
#>   raw_ntee_code subseccd has_subseccd have_filings have_extracts have_pdfs
#> 1           A20        3         TRUE           NA            NA        NA
#>      score
#> 1 11615.26
#> 
#> $num_pages
#> [1] 1
#> 
#> $cur_page
#> [1] 0
#> 
#> $page_offset
#> [1] 0
#> 
#> $per_page
#> [1] 100
#> 
#> $search_query
#> [1] "propublica"
#> 
#> $selected_state
#> NULL
#> 
#> $selected_ntee
#> [1] "NULL"
#> 
#> $selected_code
#> NULL
#> 
#> $data_source
#> [1] "ProPublica Nonprofit Explorer API: https://projects.propublica.org/nonprofits/api/\nIRS Exempt Organizations Business Master File Extract (EO BMF): https://www.irs.gov/charities-non-profits/exempt-organizations-business-master-file-extract-eo-bmf\nIRS Annual Extract of Tax-Exempt Organization Financial Data: https://www.irs.gov/uac/soi-tax-stats-annual-extract-of-tax-exempt-organization-financial-data"
#> 
#> $api_version
#> [1] 2
#> 
np_org(142007220)
#> $organization
#> $organization$id
#> [1] 142007220
#> 
#> $organization$ein
#> [1] 142007220
#> 
#> $organization$name
#> [1] "PRO PUBLICA INC"
#> 
#> $organization$careofname
#> NULL
#> 
#> $organization$address
#> [1] "155 AVE AMERICA 13 FL"
#> 
#> $organization$city
#> [1] "NEW YORK"
#> 
#> $organization$state
#> [1] "NY"
#> 
#> $organization$zipcode
#> [1] "10013-0000"
#> 
#> $organization$exemption_number
#> [1] 0
#> 
#> $organization$subsection_code
#> [1] 3
#> 
#> $organization$affiliation_code
#> [1] 3
#> 
#> $organization$classification_codes
#> [1] "1000"
#> 
#> $organization$ruling_date
#> [1] "2008-02-01"
#> 
#> $organization$deductibility_code
#> [1] 1
#> 
#> $organization$foundation_code
#> [1] 15
#> 
#> $organization$activity_codes
#> [1] "000000000"
#> 
#> $organization$organization_code
#> [1] 1
#> 
#> $organization$exempt_organization_status_code
#> [1] 1
#> 
#> $organization$tax_period
#> [1] "2019-12-01"
#> 
#> $organization$asset_code
#> [1] 9
#> 
#> $organization$income_code
#> [1] 8
#> 
#> $organization$filing_requirement_code
#> [1] 1
#> 
#> $organization$pf_filing_requirement_code
#> [1] 0
#> 
#> $organization$accounting_period
#> [1] 12
#> 
#> $organization$asset_amount
#> [1] 51252809
#> 
#> $organization$income_amount
#> [1] 39461742
#> 
#> $organization$revenue_amount
#> [1] 39246884
#> 
#> $organization$ntee_code
#> [1] "A20"
#> 
#> $organization$sort_name
#> NULL
#> 
#> $organization$created_at
#> [1] "2021-08-24T14:43:44.564Z"
#> 
#> $organization$updated_at
#> [1] "2021-08-24T14:43:44.564Z"
#> 
#> $organization$data_source
#> NULL
#> 
#> $organization$have_extracts
#> NULL
#> 
#> $organization$have_pdfs
#> NULL
#> 
#> 
#> $filings_with_data
#>   tax_prd tax_prd_yr formtype
#> 1  201912       2019        0
#> 2  201812       2018        0
#> 3  201712       2017        0
#> 4  201612       2016        0
#> 5  201512       2015        0
#> 6  201412       2014        0
#> 7  201312       2013        0
#> 8  201212       2012        0
#> 9  201112       2011        0
#>                                                                                                                               pdf_url
#> 1  https://projects.propublica.org/nonprofits/download-filing?path=02_2021_prefixes_13-20%2F142007220_201912_990_2021021917729264.pdf
#> 2 https://projects.propublica.org/nonprofits/download-filing?path=09_2019_prefixes_06-20%2F142007220_201812_990T_2019092416682802.pdf
#> 3 https://projects.propublica.org/nonprofits/download-filing?path=09_2018_prefixes_14-20%2F142007220_201712_990T_2018092415723520.pdf
#> 4                                                                                                                                <NA>
#> 5                              https://projects.propublica.org/nonprofits/download-filing?path=2016_08_T%2F14-2007220_990T_201512.pdf
#> 6                              https://projects.propublica.org/nonprofits/download-filing?path=2015_07_EO%2F14-2007220_990_201412.pdf
#> 7                              https://projects.propublica.org/nonprofits/download-filing?path=2014_08_T%2F14-2007220_990T_201312.pdf
#> 8                              https://projects.propublica.org/nonprofits/download-filing?path=2013_08_T%2F14-2007220_990T_201212.pdf
#> 9                              https://projects.propublica.org/nonprofits/download-filing?path=2012_08_EO%2F14-2007220_990_201112.pdf
#>                    updated totrevenue totfuncexpns totassetsend totliabend
#> 1 2021-08-30T19:00:19.474Z   39246884     28630586     51252809     748489
#> 2 2021-08-30T18:35:56.183Z   26685933     24104461     40988939    1094094
#> 3 2019-10-15T18:35:13.728Z   43574038     18290801     38070942     746162
#> 4 2018-12-10T23:09:19.598Z   14545521     13766881     12429526     384259
#> 5 2017-06-30T00:36:10.963Z   17046930     12461149     11552064     285437
#> 6 2017-06-30T00:20:00.529Z   10324275     11486452      6865118     184272
#> 7 2017-06-30T00:19:55.546Z   13765467     10332809      7938469      95446
#> 8 2017-06-29T23:49:47.850Z   10955982      9884450      4529443     119078
#> 9 2017-06-29T23:37:23.325Z   10142780      9651650      3591151     252318
#>   pct_compnsatncurrofcr tax_pd subseccd unrelbusinccd initiationfees
#> 1                     0     NA        3             Y              0
#> 2                     0     NA        3             Y              0
#> 3                     0     NA        3             Y              0
#> 4                     0     NA        3             Y              0
#> 5                     0     NA        3             Y              0
#> 6                     0     NA        3             Y              0
#> 7                     0     NA        3             Y              0
#> 8                     0     NA        3             Y              0
#> 9                     0     NA        3             Y              0
#>   grsrcptspublicuse grsincmembers grsincother totcntrbgfts totprgmrevnue
#> 1                 0             0           0     37366751        620950
#> 2                 0             0           0     25576127        302000
#> 3                 0             0           0     43063123        110000
#> 4                 0             0           0     13765153        340000
#> 5                 0             0           0     16882164         60000
#> 6                 0             0           0     10169976             0
#> 7                 0             0           0     13678241             0
#> 8                 0             0           0     10920019             0
#> 9                 0             0           0     10115367             0
#>   invstmntinc txexmptbndsproceeds royaltsinc grsrntsreal grsrntsprsnl
#> 1      515219                   0     196774           0            0
#> 2      288025                   0     226199           0            0
#> 3       52985                   0     165746           0            0
#> 4        2208                   0     176712           0            0
#> 5        3919                   0      70956           0            0
#> 6        4176                   0     104222           0            0
#> 7         103                   0      51278           0            0
#> 8          50                   0      17096           0            0
#> 9          83                   0       4547           0            0
#>   rntlexpnsreal rntlexpnsprsnl rntlincreal rntlincprsnl netrntlinc grsalesecur
#> 1             0              0           0            0          0      213835
#> 2             0              0           0            0          0      552029
#> 3             0              0           0            0          0      256121
#> 4             0              0           0            0          0       74091
#> 5             0              0           0            0          0       65311
#> 6             0              0           0            0          0       15500
#> 7             0              0           0            0          0      108657
#> 8             0              0           0            0          0           0
#> 9             0              0           0            0          0           0
#>   grsalesothr cstbasisecur cstbasisothr gnlsecur gnlsothr netgnls grsincfndrsng
#> 1           0       214858            0    -1023        0   -1023             0
#> 2           0       551909            0      120        0     120             0
#> 3         500       257270          885    -1149     -385   -1534             0
#> 4           0        74800            0     -709        0    -709             0
#> 5           0        65311            0        0        0       0             0
#> 6           0        15500            0        0        0       0             0
#> 7           0       108657            0        0        0       0             0
#> 8           0            0            0        0        0       0             0
#> 9           0            0            0        0        0       0             0
#>   lessdirfndrsng netincfndrsng grsincgaming lessdirgaming netincgaming
#> 1              0             0            0             0            0
#> 2              0             0            0             0            0
#> 3              0             0            0             0            0
#> 4              0             0            0             0            0
#> 5              0             0            0             0            0
#> 6              0             0            0             0            0
#> 7              0             0            0             0            0
#> 8              0             0            0             0            0
#> 9              0             0            0             0            0
#>   grsalesinvent lesscstofgoods netincsales miscrevtot11e compnsatncurrofcr
#> 1             0              0           0        548213           1229154
#> 2             0              0           0        293462           1205372
#> 3             0              0           0        183718           1198715
#> 4             0              0           0        262157           1281333
#> 5             0              0           0         29891           1272433
#> 6             0              0           0         45901           1343765
#> 7             0              0           0         35845           1338618
#> 8             0              0           0         18817                 0
#> 9             0              0           0         22783           1552988
#>   othrsalwages payrolltx profndraising txexmptbndsend secrdmrtgsend
#> 1     15804374   1230934          9600              0             0
#> 2     12893421   1008426          6225              0             0
#> 3      9839291    798346             0              0             0
#> 4      7173056    562079             0              0             0
#> 5      6628192    540502             0              0             0
#> 6      6000803    475304             0              0             0
#> 7      5360565    437285             0              0             0
#> 8            0         0             0              0             0
#> 9      4522408    331237             0              0             0
#>   unsecurednotesend retainedearnend totnetassetend nonpfrea gftgrntsrcvd170
#> 1                 0               0       50504320       07       136653318
#> 2                 0               0       39894845       07       109456543
#> 3                 0               0       37324780        7        97558657
#> 4                 0               0       12045267        7        65415553
#> 5                 0               0       11266627        7        61765767
#> 6                 0               0        6680846        7        55093004
#> 7                 0               0        7843023        7        51278007
#> 8                 0               0        4410365        7        46144525
#> 9                 0               0        3338833        7        36669506
#>   txrevnuelevied170 srvcsval170 grsinc170 grsrcptsrelated170 totgftgrntrcvd509
#> 1                 0           0   1306125            1825568                 0
#> 2                 0           0    718543            1188605                 0
#> 3                 0           0    320125             822180                 0
#> 4                 0           0    245665             585055                 0
#> 5                 0           0    156563             171500                 0
#> 6                 0           0    136349              60500                 0
#> 7                 0           0     94444                  0                 0
#> 8                 0           0     70024                  0                 0
#> 9                 0           0     41334                  0                 0
#>   grsrcptsadmissn509 txrevnuelevied509 srvcsval509 subtotsuppinc509 totsupp509
#> 1                  0                 0           0                0          0
#> 2                  0                 0           0                0          0
#> 3                  0                 0           0                0          0
#> 4                  0                 0           0                0          0
#> 5                  0                 0           0                0          0
#> 6                  0                 0           0                0          0
#> 7                  0                 0           0                0          0
#> 8                  0                 0           0                0          0
#> 9                  0                 0           0                0          0
#>         ein
#> 1 142007220
#> 2 142007220
#> 3 142007220
#> 4 142007220
#> 5 142007220
#> 6 142007220
#> 7 142007220
#> 8 142007220
#> 9 142007220
#> 
#> $filings_without_data
#>   tax_prd tax_prd_yr formtype formtype_str
#> 1  201012       2010        0          990
#> 2  200912       2009        0          990
#> 3  200812       2008        0          990
#> 4  200812       2008        0          990
#> 5  200712       2007        0          990
#> 6  200712       2007        0          990
#> 7  200712       2007        0          990
#>                                                                                                  pdf_url
#> 1 https://projects.propublica.org/nonprofits/download-filing?path=2011_08_EO%2F14-2007220_990_201012.pdf
#> 2 https://projects.propublica.org/nonprofits/download-filing?path=2010_08_EO%2F14-2007220_990_200912.pdf
#> 3 https://projects.propublica.org/nonprofits/download-filing?path=2010_08_EO%2F14-2007220_990_200812.pdf
#> 4 https://projects.propublica.org/nonprofits/download-filing?path=2009_09_EO%2F14-2007220_990_200812.pdf
#> 5 https://projects.propublica.org/nonprofits/download-filing?path=2010_08_EO%2F14-2007220_990_200712.pdf
#> 6 https://projects.propublica.org/nonprofits/download-filing?path=2008_10_EO%2F14-2007220_990_200712.pdf
#> 7 https://projects.propublica.org/nonprofits/download-filing?path=2008_08_EO%2F14-2007220_990_200712.pdf
#> 
#> $data_source
#> [1] "ProPublica Nonprofit Explorer API: https://projects.propublica.org/nonprofits/api/\nIRS Exempt Organizations Business Master File Extract (EO BMF): https://www.irs.gov/charities-non-profits/exempt-organizations-business-master-file-extract-eo-bmf\nIRS Annual Extract of Tax-Exempt Organization Financial Data: https://www.irs.gov/uac/soi-tax-stats-annual-extract-of-tax-exempt-organization-financial-data"
#> 
#> $api_version
#> [1] 2
#> 
# }